refactor: modularize Zustand store
