+++ src/MacroBuilder.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
-import { useStore, type Macro, type MacroType } from './store';
+import { useStore } from './store';
+import type { Macro, MacroType } from './store/macros';
 import { useToastStore } from './toastStore';
 import MacroInstructions from './MacroInstructions';
